Fixed logging macros to actually work!
Any variables passed to the macros were not being evaluated at all,
but were simply treated as a quoted expression.  For strings, this
was fine, but it produces confusing (and rather useless) results
when passing variables to the function.

Simply not escaping the expression helped, but then passed variables
were evaluated in the context of the Logging module, which was also
not useful.

This version requires calling `@Logging.configure(level=xxx)`,
which sets the log level, and evaluates the macros in the calling
module, which allows them to be used there as well.

Thanks to Dominique Orban (@dpo) for pointing out the main issues.